Move requests into new files
