+++ crates/bitwarden/src/auth/mod.rs
@@ -1,5 +1,4 @@
 pub(super) mod api;
 pub mod login;
-pub mod request;
 pub mod response;
 pub mod renew;
