--- crates/bitwarden/src/auth/request/mod.rs
@@ -1,5 +1,0 @@
-mod api_key_login_request;
-mod password_login_request;
-
-pub use api_key_login_request::*;
-pub use password_login_request::*;
